requirements.txt
# Django
django==4.0.1

# Database
psycopg2-binary==2.9.3

# REST API
django-cors-headers==3.11.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0

# Docs
drf-yasg==1.20.0

# Linters
flake8==4.0.1

# WSGI-server
whitenoise==5.3.0
gunicorn==20.1.0

# Tests
pytest==6.2.5
pytest-cov==3.0.0
pytest-django==4.5.2
pytest-factoryboy==2.1.0
factory-boy==3.2.1

# Codecov integration
codecov==2.1.12
